앱 추가를 추가
