+++ src/frontend/src/components/common/Error.tsx
@@ -1,6 +1,5 @@
-import { AlertCircle, RefreshCw, Home, Link as LinkIcon } from 'lucide-react';
+import { AlertCircle, Home, Link as LinkIcon, RefreshCw } from 'lucide-react';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 
 interface ErrorComponentProps {
   title: string;
@@ -20,8 +19,6 @@ export const ErrorComponent = ({
   goHome = false,
   customLink,
 }: ErrorComponentProps) => {
-  const router = useRouter();
-
   return (
     <div className="flex flex-col items-center justify-center min-h-screen gap-4">
       <div className="flex items-center gap-2 text-destructive">
@@ -37,7 +34,7 @@ export const ErrorComponent = ({
         {refresh && (
           <button
             type="button"
-            onClick={() => router.refresh()}
+            onClick={() => window.location.reload()}
             className="flex items-center px-4 py-2 text-sm text-white bg-primary rounded-md hover:bg-primary/90"
           >
             <RefreshCw className="w-4 h-4 mr-2" />
